# UT5/Supuesto práctico proyecto/TejeraSantana_Adoney_supuesto_dsw_ut5/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Supuesto UT5 Tejera Santana Adoney",
        "description": "Documentacion de supuesto ut5",
        "version": "11.31"
    },
    "servers": [
        {
            "url": "http://my-default-host.com",
            "description": "Servidor principal de la API"
        }
    ],
    "paths": {
...
        "/api/adoneytj/planes_entrenamiento": {
            "get": {
                "tags": [
                    "Planes_entrenamiento"
                ],
                "summary": "Listar planes de entrenamiento",
                "description": "Lista todos los planes de entrenamiento, solo admin y entrenadores tienen autorización",
                "operationId": "indexPlanesEntrenamiento",
                "responses": {
                    "200": {
                        "description": "Operación exitosa"
                    },
                    "401": {
                        "description": "Sin autorización, debe ser admin o entrenador"
                    }
                }
            },
            "post": {
                "tags": [
                    "Planes_entrenamiento"
                ],
                "summary": "Crea un plane de entrenamiento",
                "description": "Crea un plan de entrenamiento, solo admin y entrenadores tienen autorización",
                "operationId": "storePlanesEntrenamiento",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "id_entrenador",
                                    "id_cliente",
                                    "nombre",
                                    "fecha_inicio",
                                    "fecha_fin"
                                ],
                                "properties": {
                                    "id_entrenador": {
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "id_cliente": {
                                        "type": "integer",
                                        "example": 2
                                    },
                                    "nombre": {
                                        "type": "string",
                                        "example": "Hay que entrenar"
                                    },
                                    "fecha_inicio": {
                                        "type": "string",
                                        "format": "date",
                                        "example": "2025-2-8"
                                    },
                                    "fecha_fin": {
                                        "type": "string",
                                        "format": "date",
                                        "example": "2054-3-12"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Operación exitosa"
                    },
                    "401": {
                        "description": "Sin autorización, debe ser admin o entrenador"
                    },
                    "205": {
                        "description": "Algún parámetro inválido"
                    }
                }
            }
        },
        "/api/adoneytj/planes_entrenamiento/{id_plan}": {
            "get": {
                "tags": [
                    "Planes_entrenamiento"
                ],
                "summary": "Obtener plane de entrenamiento",
                "description": "Obtiene un plan de entrenamiento, solo admin, el mismo entrenador y el mismo cliente tienen autorización",
                "operationId": "showPlanesEntrenamiento",
                "parameters": [
                    {
                        "name": "id_plan",
                        "in": "path",
                        "description": "ID del plan",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Operación exitosa"
                    },
                    "401": {
                        "description": "Sin autorización, debe ser admin o el entrenador/cliente del plan"
                    },
                    "205": {
                        "description": "plan no encontrado"
                    }
                }
            },
            "put": {
                "tags": [
                    "Planes_entrenamiento"
                ],
                "summary": "Actualiza un plane de entrenamiento",
                "description": "Actualiza un plan de entrenamiento, solo admin y entrenadores tienen autorización",
                "operationId": "updatePlanesEntrenamiento",
                "parameters": [
                    {
                        "name": "id_plan",
                        "in": "path",
                        "description": "ID del plan",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": false,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "id_entrenador": {
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "id_cliente": {
                                        "type": "integer",
                                        "example": 2
                                    },
                                    "nombre": {
                                        "type": "string",
                                        "example": "Hay que entrenar"
                                    },
                                    "fecha_inicio": {
                                        "type": "string",
                                        "format": "date",
                                        "example": "2025-2-8"
                                    },
                                    "fecha_fin": {
                                        "type": "string",
                                        "format": "date",
                                        "example": "2054-3-12"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Operación exitosa"
                    },
                    "401": {
                        "description": "Sin autorización, debe ser admin o entrenador"
                    },
                    "205": {
                        "description": "Algún parámetro inválido o plan no encontrado"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Planes_entrenamiento"
                ],
                "summary": "Borrar plane de entrenamiento",
                "description": "Borra un plan de entrenamiento, solo admin, el mismo entrenador tienen autorización",
                "operationId": "destroyPlanesEntrenamiento",
                "parameters": [
                    {
                        "name": "id_plan",
                        "in": "path",
                        "description": "ID del plan",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Operación exitosa"
                    },
                    "401": {
                        "description": "Sin autorización, debe ser admin o el entrenador del plan"
                    },
                    "205": {
                        "description": "plan no encontrado"
                    }
                }
            }
        },
        "/api/adoneytj/planes_entrenamiento/{id_plan}/tablas_entrenamiento": {
            "post": {
                "tags": [
                    "Planes_entrenamiento"
                ],
                "summary": "Añade una tabla al plan de entrenamiento",
                "description": "Añade una tabla al plan de entrenamiento indicado, solo admin, el mismo entrenador tienen autorización",
                "operationId": "añadirTabla-PlanesEntrenamiento",
                "parameters": [
                    {
                        "name": "id_plan",
                        "in": "path",
                        "description": "ID del plan",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "id_tabla"
                                ],
                                "properties": {
                                    "id_tabla": {
                                        "type": "integer",
                                        "example": 1
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Operación exitosa"
                    },
                    "401": {
                        "description": "Sin autorización, debe ser admin o el entrenador del plan"
                    },
                    "205": {
                        "description": "plan/tabla no encontrada"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Planes_entrenamiento"
                ],
                "summary": "Elimina una tabla al plan de entrenamiento",
                "description": "Eilmina una tabla al plan de entrenamiento indicado, solo admin, el mismo entrenador tienen autorización",
                "operationId": "eliminarTabla-PlanesEntrenamiento",
                "parameters": [
                    {
                        "name": "id_plan",
                        "in": "path",
                        "description": "ID del plan",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "id_tabla"
                                ],
                                "properties": {
                                    "id_tabla": {
                                        "type": "integer",
                                        "example": 1
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Operación exitosa"
                    },
                    "401": {
                        "description": "Sin autorización, debe ser admin o el entrenador del plan"
                    },
                    "205": {
                        "description": "plan/tabla no encontrada"
                    }
                }
            }
        },
...
    },
...
}